refactor: LottoMaker 삭제
